Disable automatic checking of global variables (for now)
I've found that the naive script I wrote to 'warn about' references to
global variables can really get on your nerves because every time it
runs it resets the height of the quickfix list. Also it can give a lot
of output on bigger scripts which is so overwhelming it becomes useless.
Until I can decrease the verbosity of the script I'm disabling automatic
globals checking by default. Users can still enable it if they want and
of course you can always execute :CheckGlobals(!) manually.