Reliably fix userdata errors that have been bugging users for a while
Finally ran into this myself today, after building Vim from source and
editing a Lua script. Turns out the issue wasn't properly fixed yet! I
was able to read the updated if_lua.txt documentation now though, so I
understood the cause of the problem and how to solve it once and for
all. This should be compatible with the old Lua Interface for Vim and
the new one. See also the following issues:

 * https://github.com/xolox/vim-lua-ftplugin/pull/23
 * https://github.com/Shougo/neocomplcache.vim/issues/401