Provide option to inspect omni completion module loading list
When set, the lua_safe_omni_modules option restricts the modules
to be loaded for omni completion to the standard Lua modules.
A list of modules that would have been loaded if this option was
not set can be used to track down modules having side effects
while loading by iteratively refining lua_omni_blacklist.